NP-110 simplifying the whole process
